moved chords layout to separate class
