+++ src/engraving/libmscore/layout/layoutbeams.h
@@ -36,9 +36,11 @@ class LayoutBeams
 public:
 
     static bool isTopBeam(Ms::ChordRest* cr);
+    static bool notTopBeam(Ms::ChordRest* cr);
     static void createBeams(Ms::Score* score, LayoutContext& lc, Ms::Measure* measure);
     static void restoreBeams(Ms::Measure* m);
     static void breakCrossMeasureBeams(Ms::Measure* measure);
+    static void respace(std::vector<Ms::ChordRest*>* elements);
 
 private:
     static void beamGraceNotes(Ms::Score* score, Ms::Chord* mainNote, bool after);
