+++ src/engraving/libmscore/layout/layoutchords.h
@@ -19,8 +19,28 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
+#ifndef MU_ENGRAVING_LAYOUTCHORDS_H
+#define MU_ENGRAVING_LAYOUTCHORDS_H
 
-#ifndef __LAYOUT_H__
-#define __LAYOUT_H__
+#include <QtGlobal>
+#include <vector>
 
-#endif
+namespace Ms {
+class Score;
+class Segment;
+class Note;
+class Staff;
+}
+
+namespace mu::engraving {
+class LayoutChords
+{
+public:
+
+    static void layoutChords1(Ms::Score* score, Ms::Segment* segment, int staffIdx);
+    static qreal layoutChords2(std::vector<Ms::Note*>& notes, bool up);
+    static void layoutChords3(Ms::Score* score, std::vector<Ms::Note*>&, const Ms::Staff*, Ms::Segment*);
+};
+}
+
+#endif // MU_ENGRAVING_LAYOUTCHORDS_H
