+++ src/engraving/libmscore/layout/layouttuplets.cpp
@@ -19,8 +19,30 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
+#include "layouttuplets.h"
 
-#ifndef __LAYOUT_H__
-#define __LAYOUT_H__
+#include "../chordrest.h"
+#include "../tuplet.h"
 
-#endif
+using namespace mu::engraving;
+using namespace Ms;
+
+bool LayoutTuplets::notTopTuplet(ChordRest* cr)
+{
+    Tuplet* t = cr->tuplet();
+    if (t && t->elements().front() == cr) {
+        // find top level tuplet
+        while (t->tuplet()) {
+            t = t->tuplet();
+        }
+        // consider tuplet cross if anything moved within it
+        if (t->cross()) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    // no tuplet or not first element
+    return false;
+}
