+++ src/engraving/libmscore/libmscore.cmake
@@ -357,6 +357,10 @@ set(LIBMSCORE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/layout/layoutmeasure.h
     ${CMAKE_CURRENT_LIST_DIR}/layout/layoutbeams.cpp
     ${CMAKE_CURRENT_LIST_DIR}/layout/layoutbeams.h
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layouttuplets.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layouttuplets.h
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutchords.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/layout/layoutchords.h
     ${CMAKE_CURRENT_LIST_DIR}/layout/verticalgapdata.cpp
     ${CMAKE_CURRENT_LIST_DIR}/layout/verticalgapdata.h
 )
