+++ src/engraving/style/styledef.h
@@ -1460,6 +1460,15 @@ inline uint qHash(Sid id)
 }
 
 //---------------------------------------------------------
+//   VerticalAlignRange
+//---------------------------------------------------------
+
+enum class VerticalAlignRange {
+    SEGMENT, MEASURE, SYSTEM
+};
+
+
+//---------------------------------------------------------
 //   StyledProperty
 ///   \cond PLUGIN_API \private \endcond
 //---------------------------------------------------------
